refactor: ReadingInfoSelect 컴포넌트 분리
